Refactored `Callbacks` `impl`s into `mod callbacks`.
